Added 3D MR phantom generation. Updated README and bumped version for new release.
