+++ examples/mr_shepp_logan.py
@@ -1,13 +1,10 @@
 '''Demonstrate how to use MR Shepp-Logan.'''
 
-from phantominator import mr_shepp_logan
+from phantominator import shepp_logan
 
 if __name__ == '__main__':
 
-    N = 128
-    M0, T1, T2 = mr_shepp_logan(N, T2star=True)
-
-    print(M0.shape)
-
-    from mr_utils import view
-    view(M0, montage_axis=-1)
+    # Get proton density, T1, and T2 maps for (L, M, N) sized Shepp-
+    # Logan phantom.
+    L, M, N = 256, 255, 20
+    M0, T1, T2 = shepp_logan((L, M, N), MR=True)
