+++ phantominator/__init__.py
@@ -1,4 +1,5 @@
 '''Bring functions up to the correct level.'''
 
-from .shepp_logan import shepp_logan
+from .ct_shepp_logan import ct_shepp_logan
 from .mr_shepp_logan import mr_shepp_logan
+from .shepp_logan import shepp_logan
